demo fused
