+++ .npmignore
@@ -1,6 +1,7 @@
 node_modules
 npm-debug.log
 test
+demo
 .yalc
 yalc.lock
 .vscode
