+++ demo/index.ts
@@ -3,7 +3,7 @@ import vgl from 'vue-golden-layout'
 import 'golden-layout/src/css/goldenlayout-light-theme.css'
 Vue.use(vgl)
 
-import App from './test.vue'
+import App from './app.vue'
 import router from './router'
 
 new App({router, el: 'app'});
