--- test/tsconfig.json
@@ -1,24 +1,0 @@
-{
-  "compilerOptions": {
-	"outDir": "dist",
-	"sourceMap": true,
-	"inlineSources": true,
-	"noImplicitAny": false,
-	"module": "commonjs",
-	"moduleResolution": "node",
-	"experimentalDecorators": true,
-	"emitDecoratorMetadata": true,
-	"target": "es5",
-	"lib": [
-	  "dom",
-	  "es5",
-	  "es6"
-	],
-	"allowSyntheticDefaultImports": true,
-	"typeRoots": ["node_modules/@types"],
-	"baseUrl": "./"
-  },
-	"exclude": [
-		"**/*.spec.ts"
-	]
-}
