--- tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-    "target": "es5",
-    "sourceMap": true,
-    "inlineSources": true,
-    "jsx": "react",
-    "importHelpers": true,
-    "emitDecoratorMetadata": true,
-    "experimentalDecorators": true
-  }
-}
